changed Int to Double for rating

+++ src/main/java/data/Movie.java
@@ -3,7 +3,7 @@ package data;
 public class Movie {
 
     private String title;
-    private int rating;
+    private Double rating;
     private String poster;
     private int year;
     private String genre;
@@ -15,7 +15,7 @@ public class Movie {
     public Movie() {
     }
 
-    public Movie(String title, int rating, String poster, int year, String genre, String director, String plot, String actors, int id) {
+    public Movie(String title, Double rating, String poster, int year, String genre, String director, String plot, String actors, int id) {
         this.title = title;
         this.rating = rating;
         this.poster = poster;
@@ -35,11 +35,11 @@ public class Movie {
         this.title = title;
     }
 
-    public int getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
